Further development Added some new classes for simulation structure, filled main loop functions in MainApp class, some minor changes

+++ GlideSim/timer.cpp
@@ -104,7 +104,7 @@ void FpsCounter::frameTick()
 	LastTime = getTime();
 }
 
-float FpsCounter::getFPS()
+double FpsCounter::getFPS()
 {
 	return fps;
 }
